requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
cloud-init==23.2.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
ecdsa==0.18.0
entrypoints==0.3
exceptiongroup==1.1.3
fastapi==0.103.2
googlemaps==4.10.0
h11==0.14.0
httplib2==0.14.0
httptools==0.6.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iso8601==1.1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
opencv-python==4.10.0.84
pexpect==4.6.0
phonenumbers==8.13.20
py-firebase-dynamic-links==1.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.6.4
pydantic-core==2.16.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypika-tortoise==0.1.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==5.3.1
requests==2.31.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smsaero==1.1.0
sniffio==1.3.0
sos==4.5.6
ssh-import-id==5.10
starlette==0.27.0
systemd-python==234
tortoise-orm==0.20.0
Twisted==18.9.0
typing-extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.23.2
uvloop==0.17.0
wadllib==1.3.3
watchfiles==0.20.0
websockets==11.0.3
wsproto==1.2.0
zipp==1.0.0
zope.interface==4.7.1
